JSON writer working
